Use Visualizer and add style
